requirements.txt
Django==3.2.7
djangorestframework==3.13.1
xhtml2pdf==0.2.7
python_dotenv==0.20
psycopg2-binary==2.8.6
drf-yasg==1.20.0
django-cors-headers==3.6.0
django-simple-history==2.12.0
django-extensions==3.1.0
celery==4.4.6
django-filter==21.1
django-debug-toolbar==3.2.4
django-redis==4.12.1
pyjwt==1.7.1
python_dateutil==2.8.2